% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piragua-package.R
\docType{package}
\name{piragua-package}
\alias{piragua}
\alias{piragua-package}
\title{piragua: Streamflow and Trend Analysis Tools for the PIRAGUA Project}
\description{
Provides functions to compute hydrological indicators, perform trend analysis,
and visualize daily and aggregated streamflow data for observed and modelled
series within the PIRAGUA project. The package includes:
\itemize{
\item \code{\link{piragua_annual}} – Compute annual streamflow indicators.
\item \code{\link{piragua_monthly}} – Compute monthly streamflow indicators.
\item \code{\link{piragua_trend}} – Estimate trends using the Yue–Pilon and Sen’s slope methods.
\item \code{\link{piragua_plot}} – Generate integrated plots combining daily, monthly, and annual statistics.
}
}
\author{
\strong{Maintainer}: Santiago Beguería \email{santiago.begueria@csic.es}

}
\keyword{internal}
