% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piragua_monthly}
\alias{piragua_monthly}
\title{Computes a set of streamflow statistics at the monthly scale.}
\usage{
piragua_monthly(x, variable = "q", area = NULL)
}
\arguments{
\item{x}{A univariate \code{xts} object, organized assuming a
hydrological year.}

\item{area}{Optional. A scalar, indicating the catchment area in km^2.
If an area is provided, the resulting values are given in
mm (l m-2); otherwise, in m^3 s^-1. Applies only to
streamflow data (\code{var}='q').}

\item{var}{String. Name of the variable to analyise: one of 'q' (streamflow),
'pcp' (precipitation), 'tmp' (temperature).}
}
\value{
A multivariate \code{xts} object with the following columns:
\describe{
\item{year}{the year for which the statistics are computed; note that hydrological years
(from previous october to september) are used}
\item{month}{the month for which the statistics are computed}
\item{mean}{mean daily streamflow (m3 s-1 or mm)}
\item{max}{maximum daily value}
\item{min}{minimum daily value}
\item{sd}{standard deviation of daily streamflow (m3 s-1 or mm)}
\item{q10}{10th percentile of daily streamflow (m3 s-1 or mm)}
\item{q25}{25th percentile of daily streamflow (m3 s-1 or mm)}
\item{q50}{50th percentile of daily streamflow (median) (m3 s-1 or mm)}
\item{q75}{75th percentile of daily streamflow (m3 s-1 or mm)}
\item{q90}{90th percentile of daily streamflow (m3 s-1 or mm)}
\item{iqr}{inter-quantile range (q75-q25) (m3 s-1 or mm)}
\item{uqr}{upper-quartile range (q75-q50) (m3 s-1 / mm)}
\item{lqr}{lower-quartile range (q50-q25) (m3 s-1 / mm)}
\item{qsk}{quantile skewness, >0 for right skewness ((uqr-lqr)/iqr) (-) \link{-1, 1}}
\item{dku}{decile skewness, >0 for right skewness ((q90-q50)-(q50-q10))/(q90-q10)) (-) \link{-1, 1}}
\item{n_q10}{number of days below the 10th percentile of the whole series}
\item{n_q25}{number of days below the 25th percentile of the whole series}
\item{n_q50}{number of days above the 50th percentile of the whole series}
\item{n_q75}{number of days above the 75th percentile of the whole series}
\item{n_q90}{number of days above the 90th percentile of the whole series}
\item{v_q10}{volume of steamflow below the 10th percentile of the whole series (hm3 or mm)}
\item{v_q25}{volume of steamflow below the 25th percentile of the whole series (hm3 or mm)}
\item{v_q50}{volume of steamflow above the 50th percentile of the whole series (hm3 or mm)}
\item{v_q75}{volume of steamflow above the 75th percentile of the whole series (hm3 or mm)}
\item{v_q90}{volume of steamflow above the 90th percentile of the whole series (hm3 or mm)}
\item{n}{number of days with data}
\item{na}{number of days with missing data}
\item{na_3}{number of periods of 3 consecutive days with missing data}
For variable \code{q} the following indices are also computed:
#' \describe{
\item{vcn3}{baseflow index: daily minimum streamflow averaged over 3 consecutive days (m3 s-1 / mm)}
\item{vcn7}{baseflow index: daily minimum streamflow averaged over 3 consecutive days (m3 s-1 / mm)}
}
For variable \code{pcp} the following columns are also included:
#' \describe{
\item{nrainy}{number of rainy days}
\item{total}{total rain (mm)}
\item{meanrainy}{mean precipitation of rainy days (mm / day)}
\item{sdrainy}{standard deviation of precipitation of rainy days (mm / day)}
}
}
\description{
Computes a set of streamflow statistics at the monthly scale.
}
\section{References}{
NA
}

\examples{

# Example 1 (not run)
# monthly <- piragua_monthly(dat)

}
