% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piragua_plot}
\alias{piragua_plot}
\title{Produces a plot with time series of all the statistics, for a given station.}
\usage{
piragua_plot(da, mo, an, tr, co, st, fo = NA, variable = "q", save = TRUE)
}
\arguments{
\item{da}{A univariate \code{xts} object with the daily streamflow on a given
station.}

\item{mo}{A data frame with monthly streamflow statistics for the
same station, typically from a call to \code{piragua_monthly()}.}

\item{an}{A data frame with annual streamflow statistics for the
same station, typically from a call to \code{piragua_annual()}.}

\item{st}{The number of the station.}

\item{fo}{Character: a suffix used to construct the names of resulting,
files e.g. '1951-2015'. Defaults to NA, so no suffix is used.}

\item{save}{Logical: whether or not to write the results to disk. If FALSE
it will not save to disk, but a plot will be returned. Defaults
to TRUE.}

\item{metadat}{Character: one of (\code{streamflow}, \code{climate}), determining
which metadata file to use. This shall depend on the type of
variable being analysed (i.e., streamflow or a climatic one).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A `ggplot` object. It also stores the plot in a file,
        if `save` is set to TRUE.
}\if{html}{\out{</div>}}
}
\description{
Produces a plot with time series of all the statistics, for a given station.
}
\section{References}{
NA
}

\examples{

# Example 1 (not run)

}
