% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{piragua_trend}
\alias{piragua_trend}
\title{Compute a linear trend following the Yue-Pilon method of prewhitening
and Sen's slope, and use a Kendall test for significance.
Using function \code{zyp.yuepilon()} from package \code{zyp.t}.}
\usage{
piragua_trend(x, variable = "q")
}
\arguments{
\item{x}{A univariate \code{xts} object with the daily streamflow on a given
station.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame object with the following columns:
}\if{html}{\out{</div>}}

\describe{
\item{lbound}{the lower bound of the trend's confidence interval.}
\item{trend}{change per year (i.e. the Sen's slope).}
\item{trendd}{change per decade.}
\item{trendp}{change over the whole period.}
\item{trend_r}{change per year, expressed in \%.}
\item{trendd_r}{change per decade, expressed in \%.}
\item{trendp_r}{change over the whole period, expressed in \%.}
\item{ubound}{upper bound of the trend's confidence interval.}
\item{tau}{Kendall's tau statistic computed on the final detrended timeseries.}
\item{sig}{Kendall's P-value computed for the final detrended timeseries.}
\item{nruns}{number of runs required to converge upon a trend.}
\item{autocor}{autocorrelation of the final detrended timeseries.}
\item{valid_frac}{fraction of the data which is valid (not NA) once autocorrelation is removed.}
\item{linear}{least squares fitted trend on the same data.}
\item{intercept}{intercept of the Sen's slope (trend).}
}
\description{
These statistics are computed for each of the variables resulting from
calls to \code{piragua_monthly()} and \code{piragua_annual()}.
}
\examples{

# Example 1 (not run)
# piragua_trend(annual)

# Example 2 (not run)
# piragua_trend(monthly)

}
\references{
Yue, S., P. Pilon, B. Phinney and G. Cavadias. The influence of
autocorrelation on the ability to detect trend in hydrological series.
\emph{Hydrological Processes} 16: 1807-1829 (2002).

David Bronaugh and Arelia Werner for the Pacific Climate Impacts
Consortium. \code{zyp}: Zhang + Yue-Pilon Trends Package. R package
version 0.10-1.1 (2019). https://CRAN.R-project.org/package=zyp.
}
